ref: move EncryptSubmissionBody into API types
