+++ src/app/modules/submission/encrypt-submission/encrypt-submission.types.ts
@@ -1,37 +1,17 @@
-import { FieldResponse } from 'src/types'
-
+import { EncryptedAttachments } from '../../../../types/api'
 import { ProcessedFieldResponse } from '../submission.types'
 
-export type EncryptSubmissionBody = {
-  responses: FieldResponse[]
-  encryptedContent: string
-  attachments?: {
-    encryptedFile?: {
-      binary: string
-      nonce: string
-      submissionPublicKey: string
-    }
-  }
-  isPreview: boolean
-  version: number
-}
-
-type Attachments = {
-  encryptedFile?: {
-    binary: string
-    nonce: string
-    submissionPublicKey: string
-  }
-}
 export type EncryptSubmissionBodyAfterProcess = {
   encryptedContent: string
-  attachments?: Attachments
+  attachments?: EncryptedAttachments
   isPreview: boolean
   version: number
   parsedResponses: ProcessedFieldResponse[]
 }
 
-export type WithAttachmentsData<T> = T & { attachmentData: Attachments }
+export type WithAttachmentsData<T> = T & {
+  attachmentData: EncryptedAttachments
+}
 
 export type WithFormData<T> = T & { formData: string }
