+++ src/types/api/encrypt-submission.ts
@@ -1,0 +1,21 @@
+import { FieldResponse } from '../response'
+
+export type EncryptSubmissionBody = {
+  responses: FieldResponse[]
+  encryptedContent: string
+  attachments?: EncryptedAttachments
+  isPreview: boolean
+  version: number
+}
+
+export type EncryptedAttachments = {
+  [fieldId: string]: {
+    encryptedFile:
+      | {
+          binary: string
+          nonce: string
+          submissionPublicKey: string
+        }
+      | undefined
+  }
+}
