+++ src/types/api/index.ts
@@ -2,3 +2,4 @@ export * from './core'
 export * from './form'
 export * from './billing'
 export * from './examples'
+export * from './encrypt-submission'
